requirements.txt
Flask
PyPDF2
scikit-learn
pandas
python-docx
numpy
nltk